Migrating over from Branch
Because the branch and master have different git sources, I was not able to merge the history from both, so I migrated the smaller repository into the bigger one manually, losing the history for the index.html file and Remind_Math_Design_Document.pdf file.  I have maintained the branch (gh-branch) in case the history is needed.